backend/requirements.txt
# Web Framework
fastapi>=0.104.1
uvicorn[standard]>=0.24.0

# Database
# SQLAlchemy 2.0+ for modern ORM with async support
sqlalchemy>=2.0.23
# PostgreSQL adapter - binary version includes compiled C extensions
# Compatible with PostgreSQL 12-16 and SQLAlchemy 2.0+
# Note: For Python 3.13, use version 2.9.11 or higher
psycopg2-binary>=2.9.11

# Data Validation
# Updated versions for Python 3.13 compatibility
pydantic>=2.9.0
pydantic-settings>=2.6.0
email-validator>=2.1.0.post1

# Authentication & Security
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
python-multipart>=0.0.6

# Task Queue
celery>=5.3.4
redis>=5.0.1

# Docker Integration
docker>=6.1.3

# Environment Configuration
python-dotenv>=1.0.0
